refactor: courseWithSecitonsをtypesから参照する
