+++ frontend/src/features/course/index.ts
@@ -2,4 +2,3 @@ export * from './components/CreateCourseButton'
 export * from './components/UpdateCourseButton'
 export * from './components/CourseTable'
 export * from './hooks/useCourse'
-export * from './types'
