+++ frontend/src/features/section/useSection.ts
@@ -2,10 +2,10 @@ import { Section } from '@prisma/client'
 import { useCallback } from 'react'
 
 import { useGetApi } from '@/hooks/useApi'
+import { CourseWithSections } from '@/types'
 import { deleteApi, postApi, putApi } from '@/utils/api'
 
 import { SectionFormRequest, SectionWithRelation } from './types'
-import { CourseWithSections } from '../course'
 
 // 作成
 // Section作成、CourseにsectionId追加（更新）
